fix: support windows compiler
